requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
cloudscraper==1.2.42
discord.py==1.3.3
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pylint==2.5.3
pyparsing==2.4.7
python-anticaptcha==0.7.0
requests==2.24.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.9
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
